Move and rename get_label_array to prepare.partition.create_partition_labels
