+++ docs/api/prepare.rst
@@ -56,3 +56,5 @@ Prepare model input
     cleanup_riv
     cleanup_wel
     cleanup_wel_layered
+
+    create_partition_labels
