+++ examples/mf6/hondsrug_partitioning.py
@@ -16,7 +16,7 @@ comparison.
 import matplotlib.pyplot as plt
 
 import imod
-from imod.mf6.multimodel.partition_generator import get_label_array
+from imod.prepare.partition import create_partition_labels
 
 # %%
 # Obtain the simulation, write it, run it, and plot some heads.
@@ -47,7 +47,7 @@ ax.set_title("hondsrug original ")
 # Now we partition the Hondsrug model
 idomain = gwf_simulation["GWF"].domain
 number_partitions = 16
-submodel_labels = get_label_array(gwf_simulation, number_partitions)
+submodel_labels = create_partition_labels(gwf_simulation, number_partitions)
 
 # %%
 # plot the partitioning array. It shows how the model will be partitioned.
