+++ examples/mf6/transport_2d.py
@@ -24,7 +24,7 @@ import pandas as pd
 import xarray as xr
 
 import imod
-from imod.mf6.multimodel.partition_generator import get_label_array
+from imod.prepare.partition import create_partition_labels
 from imod.typing.grid import nan_like, zeros_like
 
 # %%
@@ -188,7 +188,7 @@ simulation.write(modeldir, binary=False)
 # To split the model in 4 partitions, we must create a label array.
 # We use the utility function  ``get_label_array`` for that.
 
-label_array = get_label_array(simulation, 4)
+label_array = create_partition_labels(simulation, 4)
 fig, ax = plt.subplots()
 label_array.plot(ax=ax)
